convert ObservationDetails to TS
also removes the QuestionNavBar component, which was unused
